bugfix: discard stale info from search threads
has always been present it seems, it's just a race condition that doesn't really trigger if you don't start a new search without explicitly stopping the old one. In that case an old bestmove could be saved, which could decrease playing strength or crash the engine. We add a message with the hash of the current position to verify that it is the one we are currently searching, discarding the message otherwise.