tm: don't search too long on mates (nonregr)
the idea is that the first subtree is typically rather small for mating
lines. This is mostly an aesthetic change; if a mating move is on the
board the game is won anyway but let's not make our opponent wait too
long for their invevitable demise.

Elo   | 0.19 +- 3.21 (95%)
SPRT  | 20.0+0.20s Threads=1 Hash=64MB
LLR   | 1.53 (-2.08, 1.50) [-5.00, 0.00]
Games | N: 12964 W: 2715 L: 2708 D: 7541
Penta | [88, 1507, 3293, 1498, 96]
https://synthetica.eu.pythonanywhere.com/test/69/

bench: 5975772